+++ src/main/java/com/jm/online_store/controller/simple/AuthorityController.java
@@ -5,21 +5,16 @@ import com.jm.online_store.service.interf.RoleService;
 import com.jm.online_store.service.interf.UserService;
 import lombok.RequiredArgsConstructor;
 import org.springframework.security.core.Authentication;
-import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.security.crypto.bcrypt.BCryptPasswordEncoder;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
-import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestParam;
-import org.springframework.web.bind.annotation.ResponseBody;
-import org.springframework.web.multipart.MultipartFile;
 
 import javax.servlet.http.HttpServletRequest;
-import java.io.IOException;
 
 @Controller
 @RequestMapping("/authority")
@@ -68,20 +63,6 @@ public class AuthorityController {
             model.addAttribute("message", "Pls, double check previous password!");
         }
         return "redirect:/authority/profile";
-    }
-
-    @PostMapping("/uploadImage")
-    @ResponseBody
-    public String handleImagePost(@RequestParam("imageFile") MultipartFile imageFile) throws IOException {
-        User userDetails = (User) SecurityContextHolder.getContext().getAuthentication().getPrincipal();
-        return userService.updateUserImage(userDetails.getId(), imageFile);
-    }
-
-    @DeleteMapping("/deleteImage")
-    @ResponseBody
-    public String deleteImage() throws IOException {
-        User userDetails = (User) SecurityContextHolder.getContext().getAuthentication().getPrincipal();
-        return userService.deleteUserImage(userDetails.getId());
     }
 
     @GetMapping("/activatenewmail/{token}")
