+++ src/main/resources/templates/adminPage.html
@@ -55,6 +55,7 @@
             <a class="navbar-brand"
                th:each="authority : ${#authentication.getAuthorities()}"
                th:text="${authority.getAuthority()} + ' '"></a>
+            <a th:href="@{/authority/profile}" class="btn btn-primary ml-auto">Profile</a>
             <a th:href="@{/logout}" class="btn btn-warning ml-auto">Logout</a>
         </nav>
         <div class="card" >
